rename Tooltip to TooltipTrigger and now it support custom icon
